Feat: Add isImpermanenceAvailable Flag for Optional Configuration
The isImpermanenceAvailable flag allows for specific hosts to utilise
impermanence, and others to not. This flag, able to be checked in user
configurations, allows for users to define which of their directories
are needed to be maintained. A specific base-template is available in
each user's host config default.nix. This precludes specific basic
directories for getting forgotten and subsequently erased.